+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog Mapgl
 
 ## 2.7.2*
+- dashed style for edges
 - fix: Fit-to-data Max zoom range 0-18 for both graph/geo
 
 ## 2.7.1